backend/identityservice/requirements.txt
# CartNudge.ai Identity Service
# ==============================

# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# MongoDB
motor==3.3.2
pymongo==4.6.1

# Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.2

# Validation & Settings
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client
httpx==0.26.0

# Utilities
python-dotenv==1.0.1
structlog==24.1.0
orjson==3.9.13

# Development
pytest==8.3.4
pytest-asyncio==0.24.0
black==24.10.0
ruff==0.8.2

